pyproject.toml
[tool.poetry]
name = "repl_python3_EoDD-Discord-chatbot"
version = "0.1.0"
description = ""
authors = ["RosalynPearson <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.1"
aiohttp = "^3.7.4"
discord = "^1.0.1"
loguru = "^0.5.3"
munch = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
